salesManagement/requirements.txt
Django==3.0.1
djangorestframework==3.10.2
djangorestframework-datatables==0.5.0
gunicorn==19.9.0
psycopg2-binary==2.8.3
Pillow==6.2.1
